+++ src/alien.py
@@ -1,5 +1,7 @@
 import pygame
 from pygame.sprite import Sprite
+
+
 class Alien(Sprite):
 	"""A class to represent a single alien in the fleet."""
 	def __init__(self,ai_game):
@@ -27,4 +29,3 @@ class Alien(Sprite):
 		"""move the alien right or left."""
 		self.x += (self.settings.alien_speed * self.settings.fleet_direction)
 		self.rect.x = self.x
-